+++ src/main/java/ru/irinaKuznetsovaa/telegramBot/Portfolio.java
@@ -1,24 +1,23 @@
 package ru.irinaKuznetsovaa.telegramBot;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class Portfolio {
     private List<Assets> listAssets;
-    private float freeMoney;
+    private Double freeMoney;
 
-    public Portfolio(List<Assets> listAssets,float freeMoney){
+    public Portfolio(List<Assets> listAssets,Double freeMoney){
        this.listAssets=listAssets;
        this.freeMoney=freeMoney;
     }
 
-    public float getFreeMoney(){
+    public Double getFreeMoney(){
         return freeMoney;
     }
 
     public List<Assets> getListAssets() {
         return listAssets;
     }
-
-
 
 }
